+++ src/main/java/cn/edu/scau/employee/entity/Role.java
@@ -6,15 +6,14 @@ import java.io.Serializable;
 import java.util.Date;
 
 /**
- *
  * @author chen
  * @description 角色实体类
  * @date 2019/11/16
  */
 @Data
-public class Role{
+public class Role {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -23,6 +22,5 @@ public class Role{
     private Date updateDate;
 
     private String remark;
-
 
 }
